requirements.txt
astroid==2.0.4
discord.py==1.4.1
Django==2.0
google-api-python-client==1.11.0
prospector==1.1.5
pylint==2.1.1
pylint-django==2.0.2
redis==3.2.1
